requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aenum==3.1.17
    # via line-bot-sdk
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.5
    # via line-bot-sdk
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   starlette
attrs==26.1.0
    # via aiohttp
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.3
    # via uvicorn
cryptography==46.0.4
    # via google-auth
deprecated==1.3.1
    # via line-bot-sdk
distro==1.9.0
    # via google-genai
fastapi==0.115.12
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
future==1.0.0
    # via line-bot-sdk
google-auth==2.48.0
    # via google-genai
google-genai==1.61.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
line-bot-sdk==3.21.0
    # via -r requirements.in
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   google-genai
    #   line-bot-sdk
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via line-bot-sdk
pyyaml==6.0.3
    # via -r requirements.in
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
    #   line-bot-sdk
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via google-genai
starlette==0.46.2
    # via fastapi
tenacity==9.1.2
    # via google-genai
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   fastapi
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via
    #   line-bot-sdk
    #   requests
uvicorn==0.27.1
    # via -r requirements.in
websockets==15.0.1
    # via google-genai
wrapt==2.1.2
    # via deprecated
yarl==1.23.0
    # via aiohttp
